o = 93/144

+++ main_ft_printf.c
@@ -54,7 +54,7 @@ void ft_analise_types(char *format, char *result, va_list ap, t_prinlist *lst) /
             ft_add_hex_str(&result, va_arg(ap, int), lst, *format);
     }
     else if(*format == 'o')
-        ft_add_octal(&result, va_arg(ap, int), lst);
+        ft_add_octal(&result, va_arg(ap, long long), lst);
     else if(*format == 'u')
     {
         if (lst->modifier == L_ONE)//позже убрать второй Л мы используем только 1
